backend funcional

+++ backend/controllers/tareasControlers.js
@@ -4,7 +4,7 @@ const Tarea = require ('../models/tareasModel')
 
 const getTareas = asyncHandler( async(req, res) => {
 
-    const tareas = await Tarea.find()
+    const tareas = await Tarea.find({user: req.user.id})
     res.status(200).json(tareas)
 })
 
@@ -15,6 +15,7 @@ const setTarea = asyncHandler( async(req, res) =>{
     } 
     const tarea = Tarea.create({
         texto: req.body.texto,
+        user: req.user.id
     })
     res.status(201).json(tarea)
     
@@ -26,8 +27,13 @@ const updateTarea = asyncHandler( async(req, res) =>{
         res.status(404)
         throw new Error ('Tarea no encontrada')
     }
-    const updatedTarea = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new: true})
-    res.status(200).json(updatedTarea)
+    if(tarea.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error ('Usuario No autorizado')
+    }else{
+        const updatedTarea = await Tarea.findByIdAndUpdate(req.params.id, req.body, {new: true})
+        res.status(200).json(updatedTarea)
+    }
 })
 
 const deleteTarea = asyncHandler( async(req, res) =>{
@@ -36,8 +42,13 @@ const deleteTarea = asyncHandler( async(req, res) =>{
         res.status(404)
         throw new Error ('Tarea no encontrada')
     }
-    tarea.deleteOne()
-    res.status(200).json({id: tarea._id})
+    if(tarea.user.toString() !== req.user.id){
+        res.status(401)
+        throw new Error ('Usuario no autorizado')
+    }else{
+        tarea.deleteOne()
+        res.status(200).json({id: tarea._id})
+    }
 })
 
 module.exports = {
